Move ``mk_mem_initializer_cpp_internal()`` from ``mk_util.py`` to ``mk_genfile_common.py`` and adapt ``mk_util.py`` and ``mk_mem_initializer_cpp.py`` to use the code at its new location. The interface has been changed slightly so that ``mk_mem_initializer_cpp_internal()`` now returns the path the generated file. The motivation behind this is so that clients of the function know the path of the generated file.
Whilst I'm here reindent ``mk_mem_initializer_cpp.py`` and the relevant
code in ``mk_util.py``.

The purpose of this change is to have Python code common to the Python
and CMake build systems separate from Python code that is only used for
the Python build system.
